New fucntion to calculate volumes
Calculate the volume of one or more images in native space. In the default setting it will count the number of voxels (1s) of a mask e.g. c_<name>_pcereb or any other binary image. It calculate the volume in mm3 based in the affine matrix information.
In the Atlas option it calculate the volume of each of the lobuli of a resliced atlas obtained by using suit_reslice_dartel_inv